+++ CHANGELOG.md
@@ -1,6 +1,62 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.1] (April 3, 2024)
+
+**Bugfixes**
+* Fix errors when URL parameters like "filter[status]" aren't strings ([#1814])
+
+## [4.0.0] (March 7, 2024)
+
+(this space intentionally left blank)
+
+## [4.0.0-rc.1] (March 4, 2024)
+
+(this space intentionally left blank)
+
+## [4.0.0-beta.3] (February 28, 2024)
+
+**Features**
+* Add a 'gp_before_translation_table' hook ([#1792])
+
+**Bugfixes**
+
+* Add trailing slash to current_url to fix matching URLs without trailing slash ([#1785])
+* Breadcrumbs: Improve consistency across all content types ([#1789])
+* Glossary: Match single word entries of parts of speech that have no suffix rules. ([#1791])
+* Fix redirecting URL with query args ([#1797])
+* Add the breadcrumbs for the "New project" actions ([#1800])
+* Make the glossary regex more deterministic ([#1801])
+
+## [4.0.0-beta.2] (January 26, 2024)
+
+**Bugfixes**
+
+* Solve some warnings with the glossary suffixes ([#1779])
+* Remove Tour code in favor of using the dedicated Tour plugin ([#1745])
+
+## [4.0.0-beta.1] (January 25, 2024)
+
+**Features**
+* Document how to use wp-env to run tests ([#1776])
+* Improve placeholders match and visibility ([#1620])
+* Change file extension for PHP format ([#1774])
+* Set the slug ([#1768])
+* Update readme.txt ([#1773])
+* Add Plugin Preview to GlotPress ([#1748])
+* Make it possible to run tests in PHP 8 ([#1760])
+* Add suffixes for matching glossary terms ([#1373])
+* Add new PHP format ([#1626])
+
+**Bugfixes**
+* Glossary: Fix fatal array in PHP8 for invalid post data ([#1750])
+* Fix glossary matching left bounded by placeholder. ([#1733])
+* Remove unused files ([#1723])
+* Fix fatal error in PHP8.x for invalid get data ([#1754])
+
+**Locales**
+* Add a new locale: Andalusian (Andalûh) ([#1729])
+
 ## [4.0.0-alpha.11] (October 19, 2023)
 
 **Bugfixes**
@@ -598,6 +654,13 @@ Thanks to all the contributors so far: Aki Björklund, Daisuke Takahashi, Domini
 
 * Initial release.
 
+[Unreleased]: https://github.com/GlotPress/GlotPress/compare/4.0.1...HEAD
+[4.0.1]: https://github.com/GlotPress/GlotPress/compare/4.0.0...4.0.1
+[4.0.0]: https://github.com/GlotPress/GlotPress/compare/4.0.0-rc.1...4.0.0
+[4.0.0-rc.1]: https://github.com/GlotPress/GlotPress/compare/4.0.0-beta.3...4.0.0-rc.1
+[4.0.0-beta.3]: https://github.com/GlotPress/GlotPress/compare/4.0.0-beta.2...4.0.0-beta.3
+[4.0.0-beta.2]: https://github.com/GlotPress/GlotPress/compare/4.0.0-beta.1...4.0.0-beta.2
+[4.0.0-beta.1]: https://github.com/GlotPress/GlotPress/compare/4.0.0-alpha.11...4.0.0-beta.1
 [4.0.0-alpha.11]: https://github.com/GlotPress/GlotPress/compare/4.0.0-alpha.10...4.0.0-alpha.11
 [4.0.0-alpha.10]: https://github.com/GlotPress/GlotPress/compare/4.0.0-alpha.9...4.0.0-alpha.10
 [4.0.0-alpha.9]: https://github.com/GlotPress/GlotPress/compare/4.0.0-alpha.8...4.0.0-alpha.9
@@ -732,3 +795,51 @@ Thanks to all the contributors so far: Aki Björklund, Daisuke Takahashi, Domini
 [#1713]: https://github.com/GlotPress/GlotPress/pull/1713
 [#1714]: https://github.com/GlotPress/GlotPress/pull/1714
 [#1726]: https://github.com/GlotPress/GlotPress/pull/1726
+[#1776]: https://github.com/GlotPress/GlotPress/pull/1776
+[#1754]: https://github.com/GlotPress/GlotPress/pull/1754
+[#1620]: https://github.com/GlotPress/GlotPress/pull/1620
+[#1772]: https://github.com/GlotPress/GlotPress/pull/1772
+[#1771]: https://github.com/GlotPress/GlotPress/pull/1771
+[#1774]: https://github.com/GlotPress/GlotPress/pull/1774
+[#1768]: https://github.com/GlotPress/GlotPress/pull/1768
+[#1773]: https://github.com/GlotPress/GlotPress/pull/1773
+[#1723]: https://github.com/GlotPress/GlotPress/pull/1723
+[#1748]: https://github.com/GlotPress/GlotPress/pull/1748
+[#1761]: https://github.com/GlotPress/GlotPress/pull/1761
+[#1760]: https://github.com/GlotPress/GlotPress/pull/1760
+[#1733]: https://github.com/GlotPress/GlotPress/pull/1733
+[#1756]: https://github.com/GlotPress/GlotPress/pull/1756
+[#1373]: https://github.com/GlotPress/GlotPress/pull/1373
+[#1626]: https://github.com/GlotPress/GlotPress/pull/1626
+[#1736]: https://github.com/GlotPress/GlotPress/pull/1736
+[#1689]: https://github.com/GlotPress/GlotPress/pull/1689
+[#1747]: https://github.com/GlotPress/GlotPress/pull/1747
+[#1750]: https://github.com/GlotPress/GlotPress/pull/1750
+[#1752]: https://github.com/GlotPress/GlotPress/pull/1752
+[#1734]: https://github.com/GlotPress/GlotPress/pull/1734
+[#1688]: https://github.com/GlotPress/GlotPress/pull/1688
+[#1731]: https://github.com/GlotPress/GlotPress/pull/1731
+[#1730]: https://github.com/GlotPress/GlotPress/pull/1730
+[#1729]: https://github.com/GlotPress/GlotPress/pull/1729
+[#1779]: https://github.com/GlotPress/GlotPress/pull/1779
+[#1745]: https://github.com/GlotPress/GlotPress/pull/1745
+[#1801]: https://github.com/GlotPress/GlotPress/pull/1801
+[#1798]: https://github.com/GlotPress/GlotPress/pull/1798
+[#1800]: https://github.com/GlotPress/GlotPress/pull/1800
+[#1797]: https://github.com/GlotPress/GlotPress/pull/1797
+[#1785]: https://github.com/GlotPress/GlotPress/pull/1785
+[#1791]: https://github.com/GlotPress/GlotPress/pull/1791
+[#1789]: https://github.com/GlotPress/GlotPress/pull/1789
+[#1796]: https://github.com/GlotPress/GlotPress/pull/1796
+[#1786]: https://github.com/GlotPress/GlotPress/pull/1786
+[#1792]: https://github.com/GlotPress/GlotPress/pull/1792
+[#1803]: https://github.com/GlotPress/GlotPress/pull/1803
+[#1814]: https://github.com/GlotPress/GlotPress/pull/1814
+[#1815]: https://github.com/GlotPress/GlotPress/pull/1815
+[#1820]: https://github.com/GlotPress/GlotPress/pull/1820
+[#1821]: https://github.com/GlotPress/GlotPress/pull/1821
+[#1819]: https://github.com/GlotPress/GlotPress/pull/1819
+[#1818]: https://github.com/GlotPress/GlotPress/pull/1818
+[#1809]: https://github.com/GlotPress/GlotPress/pull/1809
+[#1807]: https://github.com/GlotPress/GlotPress/pull/1807
+[#1773]: https://github.com/GlotPress/GlotPress/pull/1773